requirements.txt
# conda install pytorch==1.12.1 torchvision==0.13.1 torchaudio==0.12.1 cudatoolkit=11.3 -c pytorch
transformers == 4.31.0
wandb == 0.15.5
pypianoroll == 1.0.4
librosa
pretty-midi == 0.2.10
tqdm == 4.65.0
torchtoolkit == 0.0.4